Allow arch-qualifiers in package relationships
Any relationship is permitted and the relationship is not checked for validity
just as the rest of the package relation is not checked. The set of qualifiers
currently accepted in the archive includes :any but dpkg permits even more. There are plans to use :native and real arch-names (:i386, :amd64 etc) in the future.