Add support for build-profiles
Parse the <profile.name> syntax, creating a list of profiles and whether
they are positive or negative specifications in a similar same style to the
architecture restriction list.

The syntactic correctness of the profile name is not enforced.

    https://wiki.debian.org/BuildProfileSpec

The profile name is split into the namespace and name parts.
Thanks to Johannes Schauer for the suggestion.