Fix TestPkgRelations failures with python3.
The real implementation of assertWarns performs an iteration over
sys.modules.values(), which breaks because somehow the size of
sys.modules changes during the iteration.  I can't find where this
is happening (my best guess is there is some parallelism I can't find,
but even that is looking unlikely).  For now, overriding the real
implementation with our own fixes the test failure.