Add a deb822.RestrictedWrapper class.
This can be used to expose read-only access to a Deb822 instance's field
values as strings, while restricting write access to some fields, which
are exposed via properties, possibly with custom conversion to a
different data type.

Using this for new deb822-based classes will be more flexible and more
reliable than subclassing Deb822 directly.
  - Parsing and dumping will always work as long as the data is
    well-formed 822-format (even if the format-specific fields do not
    match their particular format).
  - If there is a problem converting the custom class for a given field
    to string (including validation errors), that is uncovered
    immediately when the field is set instead of when someone asks for
    the string representation of the whole object.
  - Much simpler interaction with Deb822 itself - it is only accessed
    through the collections.MutableMap interface and the 'dump' method,
    so users don't need to worry about how their custom code interacts
    with other internal methods of Deb822.

I wish we had done something like this for the current subclasses (e.g.
Packages and Sources).  It's a bit difficult to change that now, but we
can at least make new stuff wrap instead of inherit.