copyright: Add a License class.
License is an immutable (synopsis, text) tuple.  Later, we can add
methods to attempt to parse the synopsis as an expression of
alternatvies - for now it just keeps the data and provides an easy way
to generate data to shove into a License field from free-form text.

This change also makes the license field of Header actually parse to and
from the License class.