deb822.Deb822Dict.dump: Add a text_mode parameter.
This will allow dumping to file(-like) objects that assume text/unicode
input (e.g. files that are opened in text mode, and io.StringIO).

(There is a lot of encoding stuff in deb822 that makes my head hurt.
Most of it is almost certainly wrong.  Lots of it could be ripped out,
but it's hard to know which users would break as a result or how.  It's
almost tempting to rip it out anyway and just commit to helping the
reverse-dependencies properly deal with text...)