copyright: Accept (and re-write) Format-Specification.
Old versions of the format used Format-Specification instead of Format
to specify the version of the format.  This change just rewrites that
field upon import as Format.

There are other issues with old specs that this change does not attempt
to solve, including:
  - Different format for Files (e.g. comma-separated instead of
    whitespace, ** vs *, beginning patterns with ./).
  - Paragraphs that are not in the current standard (e.g.
    Removed-Files).
I think this module should just worry about making sure these old
formats actually parse.  Other tools can build on top of it to try to
normalize old field values.