changelog: Remove support for broken timezone names
The Debian changelog trailer regex intended to support a timezone name
inside parenthesis, like this:

 -- Name <mail@example.org>  Sat, 30 May 2015 03:18:43 +0200 (CEST)

is bogus, and it only accepts one character. As in:

 -- Name <mail@example.org>  Sat, 30 May 2015 03:18:43 +0200 (C)

which while also “valid”, as per RFC822, they are pretty much useless
as per RFC5322. It seems the regex was “inherited” from the dpkg
implementation which has kept that part unchanged since its inception
in dpkg 1.3.0, 1996-08.

So let's just drop it instead of fixing the regex, given that the
timezone name has never been accepted, many time-parsing functions
ignore it, it is redundant, declared obsolete by RFC5322 and Debian
policy dropped an explicit reference to it due to bug 569174.

Signed-off-by: Guillem Jover <guillem@debian.org>