Fix comment handling tests (#807351)
Newer python-apt raises exceptions when it can't parse the comments in a
deb822 file so testing for them should replace testing for a misparsed
file.

Based on patches from Neil Williams and Barry Warsaw.

Since this test now relies on features that are only in python-apt >= 1.1,
include that version requirement in the build-depends.